Improve babycloudrunner_update 96.32->96.54% (s16 cast on AirMeter table value for s16toFloat)

+++ src/main/dll/dll_014C_babycloudrunner.c
@@ -1083,7 +1083,7 @@ void babycloudrunner_update(int* obj)
                         sub->runnerState = 2;
                         GameBit_Set(0x66, 0);
                         (*gGameUIInterface)->initAirMeter(gBabyCloudRunnerAirMeterValues[sub->runnerIndex], 0x5d1);
-                        s16toFloat((int)((char*)sub + 0x238), gBabyCloudRunnerAirMeterValues[sub->runnerIndex]);
+                        s16toFloat((int)((char*)sub + 0x238), (s16)gBabyCloudRunnerAirMeterValues[sub->runnerIndex]);
                     }
                     fn_8019E3F4(obj);
                     return;
